Add documation for dealing with /dev/random issues
You may occasionally see that nodes will not immediately connect this
with the hub. In some cases the delay has been as high as 40 minutes.
This is due to a known problem with linux containers where there is a
lack of entropy within the running container. [James Bayer](http://blog.pivotal.io/cloud-foundry-pivotal/features/challenges-with-randomness-in-multi-tenant-linux-container-platforms) wrote a great article explaining the core of the problem.

Resolve #14